Move things which should not be in core/testing.go out of it
