+++ core/inbox_test.go
@@ -35,9 +35,9 @@ func TestUpdateMessagePool(t *testing.T) {
 		m := types.NewSignedMsgs(2, mockSigner)
 		requireAdd(t, ib, m[0], m[1])
 
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{})
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{})
 
-		newChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{m[1]}})
+		newChain := requireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{m[1]}})
 
 		assert.NoError(t, ib.HandleNewHead(ctx, oldChain, newChain))
 		assertPoolEquals(t, p, m[0])
@@ -54,7 +54,7 @@ func TestUpdateMessagePool(t *testing.T) {
 		m := types.NewSignedMsgs(3, mockSigner)
 		requireAdd(t, ib, m[0], m[1])
 
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{m[2]}})
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{m[2]}})
 
 		assert.NoError(t, ib.HandleNewHead(ctx, oldChain, oldChain)) // sic
 		assertPoolEquals(t, p, m[0], m[1])
@@ -71,9 +71,9 @@ func TestUpdateMessagePool(t *testing.T) {
 		m := types.NewSignedMsgs(7, mockSigner)
 		requireAdd(t, ib, m[2], m[5])
 
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{m[0], m[1]}})
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{m[0], m[1]}})
 
-		newChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent,
+		newChain := requireChainWithMessages(t, chainProvider.Builder, parent,
 			msgsSet{msgs{m[2], m[3]}},
 			msgsSet{msgs{m[4]}},
 			msgsSet{msgs{m[0]}},
@@ -96,9 +96,9 @@ func TestUpdateMessagePool(t *testing.T) {
 		m := types.NewSignedMsgs(7, mockSigner)
 		requireAdd(t, ib, m[2], m[5])
 
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{m[0]}, msgs{m[1]}})
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{m[0]}, msgs{m[1]}})
 
-		newChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent,
+		newChain := requireChainWithMessages(t, chainProvider.Builder, parent,
 			msgsSet{msgs{m[2], m[3]}},
 			msgsSet{msgs{m[4]}, msgs{m[0]}, msgs{}, msgs{}},
 			msgsSet{msgs{}, msgs{m[5], m[6]}},
@@ -119,13 +119,13 @@ func TestUpdateMessagePool(t *testing.T) {
 		m := types.NewSignedMsgs(6, mockSigner)
 		requireAdd(t, ib, m[3], m[5])
 
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent,
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent,
 			msgsSet{msgs{m[0]}},
 			msgsSet{msgs{m[1]}},
 			msgsSet{msgs{m[2]}},
 		)
 
-		newChain := core.RequireChainWithMessages(t, chainProvider.Builder, oldChain[0],
+		newChain := requireChainWithMessages(t, chainProvider.Builder, oldChain[0],
 			msgsSet{msgs{m[3]}},
 			msgsSet{msgs{m[4], m[5]}},
 		)
@@ -144,13 +144,13 @@ func TestUpdateMessagePool(t *testing.T) {
 
 		m := types.NewSignedMsgs(7, mockSigner)
 		requireAdd(t, ib, m[6])
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent,
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent,
 			msgsSet{msgs{m[0]}},
 			msgsSet{msgs{m[1]}},
 			msgsSet{msgs{m[2]}},
 		)
 
-		newChain := core.RequireChainWithMessages(t, chainProvider.Builder, oldChain[0],
+		newChain := requireChainWithMessages(t, chainProvider.Builder, oldChain[0],
 			msgsSet{msgs{m[3]}},
 			msgsSet{msgs{m[4]}},
 			msgsSet{msgs{m[5]}},
@@ -172,12 +172,12 @@ func TestUpdateMessagePool(t *testing.T) {
 		m := types.NewSignedMsgs(7, mockSigner)
 		requireAdd(t, ib, m[6])
 
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent,
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent,
 			msgsSet{msgs{m[0]}, msgs{m[1]}},
 			msgsSet{msgs{m[2]}},
 		)
 
-		newChain := core.RequireChainWithMessages(t, chainProvider.Builder, oldChain[0],
+		newChain := requireChainWithMessages(t, chainProvider.Builder, oldChain[0],
 			msgsSet{msgs{m[3]}},
 			msgsSet{msgs{m[4]}},
 			msgsSet{msgs{m[5]}, msgs{m[1], m[2]}},
@@ -198,13 +198,13 @@ func TestUpdateMessagePool(t *testing.T) {
 		m := types.NewSignedMsgs(6, mockSigner)
 		requireAdd(t, ib, m[3], m[5])
 
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent,
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent,
 			msgsSet{msgs{m[0]}},
 			msgsSet{msgs{m[1]}},
 			msgsSet{msgs{m[2]}},
 		)
 
-		newChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent,
+		newChain := requireChainWithMessages(t, chainProvider.Builder, parent,
 			msgsSet{msgs{m[0]}, msgs{m[1]}, msgs{m[2]}},
 		)
 
@@ -221,7 +221,7 @@ func TestUpdateMessagePool(t *testing.T) {
 		ib := core.NewInbox(p, 10, chainProvider, chainProvider)
 		m := types.NewSignedMsgs(4, mockSigner)
 
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent,
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent,
 			msgsSet{msgs{m[0]}},
 			msgsSet{msgs{m[1]}},
 			msgsSet{msgs{m[2]}},
@@ -243,9 +243,9 @@ func TestUpdateMessagePool(t *testing.T) {
 		m := types.NewSignedMsgs(3, mockSigner)
 		requireAdd(t, ib, m[0], m[1])
 
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{}})
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{}})
 
-		newChain := core.RequireChainWithMessages(t, chainProvider.Builder, oldChain[0], msgsSet{msgs{m[1], m[2]}})
+		newChain := requireChainWithMessages(t, chainProvider.Builder, oldChain[0], msgsSet{msgs{m[1], m[2]}})
 
 		assert.NoError(t, ib.HandleNewHead(ctx, nil, newChain))
 		assertPoolEquals(t, p, m[0])
@@ -262,12 +262,12 @@ func TestUpdateMessagePool(t *testing.T) {
 		m := types.NewSignedMsgs(7, mockSigner)
 		requireAdd(t, ib, m[2], m[5])
 
-		oldChain := core.RequireChainWithMessages(t, chainProvider.Builder, parent,
+		oldChain := requireChainWithMessages(t, chainProvider.Builder, parent,
 			msgsSet{msgs{m[0]}},
 			msgsSet{msgs{m[1]}},
 		)
 
-		newChain := core.RequireChainWithMessages(t, chainProvider.Builder, oldChain[0],
+		newChain := requireChainWithMessages(t, chainProvider.Builder, oldChain[0],
 			msgsSet{msgs{m[2], m[3]}},
 			msgsSet{msgs{m[4]}},
 			msgsSet{msgs{m[5], m[6]}},
@@ -285,7 +285,7 @@ func TestUpdateMessagePool(t *testing.T) {
 
 		m := types.NewSignedMsgs(maxAge, mockSigner)
 
-		head := core.RequireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{}})[0]
+		head := requireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{}})[0]
 
 		// Add a message at each block height until maxAge is reached.
 		for i := uint(0); i < maxAge; i++ {
@@ -295,7 +295,7 @@ func TestUpdateMessagePool(t *testing.T) {
 			requireAdd(t, ib, m[i])
 
 			// update pool with tipset that has no messages
-			next := core.RequireChainWithMessages(t, chainProvider.Builder, head, msgsSet{msgs{}})[0]
+			next := requireChainWithMessages(t, chainProvider.Builder, head, msgsSet{msgs{}})[0]
 			assert.NoError(t, ib.HandleNewHead(ctx, nil, []types.TipSet{next}))
 
 			// assert all added messages still in pool
@@ -306,12 +306,12 @@ func TestUpdateMessagePool(t *testing.T) {
 		require.Equal(t, types.Uint64(11), head.At(0).Height)
 
 		// next tipset times out first message only
-		next := core.RequireChainWithMessages(t, chainProvider.Builder, head, msgsSet{msgs{}})[0]
+		next := requireChainWithMessages(t, chainProvider.Builder, head, msgsSet{msgs{}})[0]
 		assert.NoError(t, ib.HandleNewHead(ctx, nil, []types.TipSet{next}))
 		assertPoolEquals(t, p, m[1:]...)
 
 		// adding a chain of 4 tipsets times out based on final state
-		newChain := core.RequireChainWithMessages(t, chainProvider.Builder, next,
+		newChain := requireChainWithMessages(t, chainProvider.Builder, next,
 			msgsSet{msgs{}},
 			msgsSet{msgs{}},
 			msgsSet{msgs{}},
@@ -329,7 +329,7 @@ func TestUpdateMessagePool(t *testing.T) {
 		ib := core.NewInbox(p, maxAge, chainProvider, chainProvider)
 
 		m := types.NewSignedMsgs(maxAge, mockSigner)
-		head := core.RequireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{}})[0]
+		head := requireChainWithMessages(t, chainProvider.Builder, parent, msgsSet{msgs{}})[0]
 
 		// Add a message at each block height until maxAge is reached
 		for i := uint(0); i < maxAge; i++ {
@@ -353,7 +353,7 @@ func TestUpdateMessagePool(t *testing.T) {
 		}
 
 		// next tipset times out first message only
-		next := core.RequireChainWithMessages(t, chainProvider.Builder, head, msgsSet{msgs{}})[0]
+		next := requireChainWithMessages(t, chainProvider.Builder, head, msgsSet{msgs{}})[0]
 		assert.NoError(t, ib.HandleNewHead(ctx, nil, []types.TipSet{next}))
 		assertPoolEquals(t, p, m[1:]...)
 	})
@@ -408,5 +408,39 @@ func assertPoolEquals(t *testing.T, p *core.MessagePool, expMsgs ...*types.Signe
 		if !found {
 			assert.Failf(t, "wrong messages in pool", "expMsgs %v, got msgs %v (msgs doesn't contain %v)", msgsAsString(expMsgs), msgsAsString(msgs), msgAsString(m1))
 		}
+	}
+}
+
+// requireChainWithMessages creates a chain of tipsets containing the given messages
+// using the provided chain builder.  The builder stores the chain.  Note that
+// each msgSet argument is a slice of message slices.  Each slice of slices
+// goes into a successive tipset and each subslice goes into one tipset block.
+// Precondition: the root tipset must be defined.  The chain of tipsets is
+// returned in descending height order (head-first).
+// TODO: move this onto the builder, #3110
+func requireChainWithMessages(t *testing.T, builder *chain.Builder, root types.TipSet, msgSets ...[][]*types.SignedMessage) []types.TipSet {
+	var tipSets []types.TipSet
+	parent := root
+	require.True(t, parent.Defined())
+
+	for _, tsMsgSet := range msgSets {
+		if len(tsMsgSet) == 0 {
+			parent = builder.BuildOneOn(parent, nil)
+		} else {
+			parent = builder.Build(parent, len(tsMsgSet), msgBuild(t, tsMsgSet))
+		}
+		tipSets = append(tipSets, parent)
+	}
+	chain.Reverse(tipSets)
+	return tipSets
+}
+
+// msgBuild takes in the msgSet dictating which messages go on which block of
+// a test tipset and returns a build function that adds these messages to the
+// correct block using the chain.Builder.
+func msgBuild(t *testing.T, msgSet [][]*types.SignedMessage) func(*chain.BlockBuilder, int) {
+	return func(bb *chain.BlockBuilder, i int) {
+		require.True(t, i <= len(msgSet))
+		bb.AddMessages(msgSet[i], types.EmptyReceipts(len(msgSet[i])))
 	}
 }
